requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.2
python-dotenv==0.19.2
pymysql==1.0.2
